requirements.txt
o365
python-dotenv
pyyaml